frontend-service/requirements.txt
streamlit==1.46.1
pandas==2.3.1
requests
python-dotenv==1.1.1